add viz funcs to viz module
